Add files via upload
This code takes the area of occupancy (AOO) output and ranks AOOs in order of size (1=smallest, 129=largest) for each of the 7 range size metrics used to estimate AOOs. It then scales all of the means (n=129) between 0 and 1. It then calculates mean and standard deviations for the ranks for each species. Code was written by J A Smith (contact=jennifer.smith@utsa.edu) on 10/16/2018